rename updaters files
This renames the files with updater implementations with a common format.
