+++ src/image_intel.hpp
