+++ src/image_openbmc.hpp
