+++ src/image_openpower.hpp
